🎨 Devide: layout server components
